# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LiqaaSpace",
        "description": "Documentation de l'API LiqaaSpace",
        "contact": {
            "name": "Équipe LiqaaSpace",
            "email": "contact@liqaaspace.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/salles": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Liste toutes les salles",
                "operationId": "4bcbc5cae51f9a46f269de9c3a5d2a84",
                "responses": {
                    "200": {
                        "description": "Liste des salles récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Salles"
                ],
                "summary": "Créer une nouvelle salle",
                "operationId": "beb6d7634c2b595aca41e8867c72adb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "description",
                                    "responsable_id",
                                    "location",
                                    "capacite",
                                    "status"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Salle de conférence A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Grande salle équipée d'un vidéoprojecteur"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "responsable_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bâtiment A, 1er étage"
                                    },
                                    "capacite": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salle créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/salles/{id}": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Afficher une salle spécifique",
                "operationId": "e9c6c8bc56d7ea04d6e2f926a6280550",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle récupérée avec succès"
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Salles"
                ],
                "summary": "Mettre à jour une salle",
                "operationId": "c92718071b3e8c9158972bf83c64b0e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "description",
                                    "responsable_id",
                                    "location",
                                    "capacite",
                                    "status"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Salle de conférence A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Grande salle équipée d'un vidéoprojecteur"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "salles/image.jpg"
                                        }
                                    },
                                    "responsable_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bâtiment A, 1er étage"
                                    },
                                    "capacite": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salle mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Salle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salles"
                ],
                "summary": "Supprimer une salle",
                "operationId": "0481df89b8e42e4e348bb6aa1bcc7aae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle supprimée avec succès"
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Salle": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Salle de conférence A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description de la salle"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "responsable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "location": {
                        "type": "string",
                        "example": "Bâtiment A, 1er étage"
                    },
                    "capacite": {
                        "type": "integer",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}